Refactor shell completions to use shared trait and macro
Extract common completion functionality into restate-cli-util crate:
- Create CompletionProvider trait for CLI applications
- Add completion_commands! macro for generating standard structures
- Move shared logic for shell detection, generation, and installation
- Refactor both CLI and restatectl to use shared components
- Remove duplicated code while maintaining full functionality
- All tests pass and clippy is clean

This enables easy addition of completions to any future CLI tools.
